--- src/features/budget/index.ts
@@ -1,14 +1,0 @@
-// src/features/budget/index.ts
-
-export {
-  BudgetItem,
-  BudgetPanelHeader,
-  CategoryProgressBar,
-  ExpenseTable,
-  BudgetRow,
-  ActivitiesBudgetPopup,
-} from './components';
-
-export { BudgetProvider, useBudgetContext, useBudgetSupabase } from './hooks';
-
-export type { Entry, CategoryKey } from './types';
