+++ src/features/planner/components/budget/index.ts
@@ -1,4 +1,4 @@
-// src/features/budget/components/index.ts
+// src/features/planner/components/budget/index.ts
 
 export { default as BudgetItem } from './BudgetItem';
 export { default as BudgetPanelHeader } from './BudgetPanelHeader';
