+++ src/features/planner/hooks/budget/index.ts
@@ -1,4 +1,4 @@
-// src/features/budget/hooks/index.ts
+// src/features/planner/hooks/budget/index.ts
 
 export { BudgetProvider, useBudgetContext } from './BudgetContext';
 export { useBudget, useBudgetSupabase } from './useBudgetSupabase';
