+++ src/features/planner/types/budget/index.ts
@@ -1,3 +1,3 @@
-// src/features/budget/types/index.ts
+// src/features/planner/types/budget/index.ts
 
 export type { Entry, CategoryKey } from './budget';
