+++ tests/unit/features/planner/budget/components/activities/BudgetPanelHeader.test.tsx
@@ -1,10 +1,10 @@
-// tests/unit/features/budget/components/activities/BudgetPanelHeader.test.tsx
+// tests/unit/features/planner/budget/components/activities/BudgetPanelHeader.test.tsx
 
 import React from 'react';
 import { render, screen, waitFor } from '@testing-library/react';
 import { vi } from 'vitest';
-import BudgetPanelHeader from '@/features/budget/components/BudgetPanelHeader';
-import { BudgetProvider } from '@/features/budget/hooks';
+import BudgetPanelHeader from '@/features/planner/components/budget/BudgetPanelHeader';
+import { BudgetProvider } from '@/features/planner/hooks';
 
 const mockFrom = vi.fn();
 vi.mock('@/shared/lib/supabaseClient', () => ({
